refactor: complete Phase 4 - idiomatic Go simplification
Achieved 474 line reduction through genuine simplification:

- Removed redundant error wrapping throughout codebase
- Consolidated package manager tests with shared test suite
- Simplified dotfiles validation to trust standard library
- Merged doctor command into status with --health flag
- Removed ~1,165 lines of dead code and unused packages
- Inlined single-use helper functions

All tests pass. Code is now cleaner and more maintainable.
