+++ internal/commands/helpers.go
@@ -4,11 +4,9 @@
 package commands
 
 import (
-	"context"
 	"fmt"
 	"runtime"
 	"strings"
-	"time"
 
 	"github.com/richhaase/plonk/internal/orchestrator"
 	"github.com/richhaase/plonk/internal/resources"
@@ -46,31 +44,22 @@ func getOSPackageManagerSupport() map[string]bool {
 
 // getManagerInstallSuggestion returns installation instructions for different package managers
 func getManagerInstallSuggestion(manager string) string {
-	osSpecific := getOSSpecificInstallCommand(manager, runtime.GOOS)
-	if osSpecific != "" {
-		return osSpecific
-	}
-	return "Install the required package manager or change the default manager with 'plonk config edit'"
-}
-
-// getOSSpecificInstallCommand returns OS-specific installation instructions for package managers
-func getOSSpecificInstallCommand(manager, os string) string {
 	// First check if the manager is supported on this OS
 	support := getOSPackageManagerSupport()
 	if supported, exists := support[manager]; exists && !supported {
-		return fmt.Sprintf("%s is not supported on %s", manager, os)
+		return fmt.Sprintf("%s is not supported on %s", manager, runtime.GOOS)
 	}
 
 	switch manager {
 	case "homebrew":
-		switch os {
+		switch runtime.GOOS {
 		case "darwin":
 			return "Install Homebrew: /bin/bash -c \"$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)\""
 		case "linux":
 			return "Install Homebrew on Linux: /bin/bash -c \"$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)\""
 		}
 	case "npm":
-		switch os {
+		switch runtime.GOOS {
 		case "darwin":
 			return "Install Node.js and npm: brew install node OR download from https://nodejs.org"
 		case "linux":
@@ -79,21 +68,21 @@ func getOSSpecificInstallCommand(manager, os string) string {
 	case "cargo":
 		return "Install Rust and Cargo: curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh"
 	case "gem":
-		switch os {
+		switch runtime.GOOS {
 		case "darwin":
 			return "Ruby comes pre-installed on macOS. For a newer version: brew install ruby"
 		case "linux":
 			return "Install Ruby: sudo apt-get install ruby-full OR use rbenv/rvm"
 		}
 	case "go":
-		switch os {
+		switch runtime.GOOS {
 		case "darwin":
 			return "Install Go: brew install go OR download from https://go.dev/dl/"
 		case "linux":
 			return "Install Go: Download from https://go.dev/dl/ OR sudo snap install go --classic"
 		}
 	case "pip":
-		switch os {
+		switch runtime.GOOS {
 		case "darwin":
 			return "Install Python and pip: brew install python3 OR download from https://python.org"
 		case "linux":
@@ -101,7 +90,7 @@ func getOSSpecificInstallCommand(manager, os string) string {
 		}
 	}
 
-	return ""
+	return "Install the required package manager or change the default manager with 'plonk config edit'"
 }
 
 // GetMetadataString safely extracts string metadata from operation results
@@ -113,19 +102,6 @@ func GetMetadataString(result resources.OperationResult, key string) string {
 		return value
 	}
 	return ""
-}
-
-// CreateOperationContext creates a context with timeout for batch operations
-func CreateOperationContext(timeout time.Duration) (context.Context, context.CancelFunc) {
-	return context.WithTimeout(context.Background(), timeout)
-}
-
-// CheckCancellation checks if the context has been canceled and returns appropriate error
-func CheckCancellation(ctx context.Context, domain string, operation string) error {
-	if ctx.Err() != nil {
-		return fmt.Errorf("%s %s: operation canceled or timed out: %w", operation, domain, ctx.Err())
-	}
-	return nil
 }
 
 // DetermineExitCode determines the appropriate exit code based on operation results
